requirements.txt
Django==4.2.7
djangorestframework
psycopg2==2.8.6
pillow
